+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to SimpleDisenchant will be documented in this file.
 
+## [1.1.0] - 2025-01-26
+
+### Added
+- Button in Enchanting profession UI to open SimpleDisenchant
+- Drag button to action bar creates a macro to toggle the addon
+- Frame auto-docks next to profession window when opened from there
+
+### Changed
+- New Blizzard-style portrait frame (PortraitFrameTemplate)
+- Use native Blizzard disenchant icon for better quality
+- Updated interface version for TWW 12.0.1
+
 ## [1.0.0] - 2025-01-26
 
 ### Added